Refactor and remove localStorage usage
