+++ src/pages/members/MemberSearchPage.tsx
@@ -37,6 +37,8 @@ const MemberSearchPage: React.FC = () => {
     memberSearch(query || '%', { page: nextPage - 1 ?? 0 })
   }
 
+  console.log('render')
+
   return (
     <>
       <SearchForm
